Add serial connector and live server
The serial connector performs a handshake with PICduino to establish
a connection on one thread and fills a queue with data recieved.

The live server is a HTTP server that serves localhost:8000 send off
the data from the queue directly to the browser which polls a HTTP
request every second.

Improvements can likely be made in using multiprocessing rather than
using the python threading concurrency, and using utilizing a socket
connection between server and browser rather than browser polling data
via HTTP requests.

Must still implement a way for the server/browser to deal with message
framing of the serial data and then display this data. Also work out the
details for how the GUI transceiver handshake ACK would actually look.